Remove global lock skip test with cache enable from race test
